Move success page to individual page
