+++ pages/wards/[id]/schedule-visit.js
@@ -17,6 +17,7 @@ import moment from "moment";
 import verifyToken from "../../../src/usecases/verifyToken";
 import TokenProvider from "../../../src/providers/TokenProvider";
 import LabelHeader from "../../../src/components/LabelHeader";
+import Router from "next/router";
 
 const isValidPhoneNumber = (input) => {
   const validator = PhoneNumberUtil.getInstance();
@@ -36,28 +37,6 @@ const isValidName = (input) => {
 const isValidDate = ({ year, month, day, hour, minute }) => {
   const parsed = moment({ year, month, day, hour, minute });
   return parsed.isValid() && parsed.isAfter(moment());
-};
-
-const Success = ({ id }) => {
-  return (
-    <Layout title="Schedule a virtual visit">
-      <GridRow>
-        <GridColumn width="two-thirds">
-          <Heading>Virtual visit scheduled</Heading>
-
-          <Text>
-            Your virtual visit has been scheduled and the key contact has been
-            sent an SMS with their scheduled time.
-          </Text>
-
-          <ActionLink href={`/wards/${id}/schedule-visit`}>
-            Schedule another visit
-          </ActionLink>
-          <ActionLink href={`/wards/${id}/visits`}>View visits</ActionLink>
-        </GridColumn>
-      </GridRow>
-    </Layout>
-  );
 };
 
 const CheckAnswers = ({ onSubmit, patientName, contactNumber, callTime }) => {
@@ -207,7 +186,7 @@ const Home = ({ id }) => {
       const { success, err } = await response.json();
 
       if (success) {
-        setSuccess(true);
+        Router.push(`/wards/${id}/schedule-success`);
       } else {
         console.error(err);
       }
@@ -221,10 +200,6 @@ const Home = ({ id }) => {
       }
     }
   });
-
-  if (success) {
-    return <Success id={id} />;
-  }
 
   if (confirmation) {
     return (
